Utils: Include context div helper
